Rename plugin.loader to plugin.template
- Add helper function for rendering a template
