+++ InvenTree/InvenTree/settings.py
@@ -343,7 +343,7 @@ TEMPLATES = [
             ],
             'loaders': [(
                 'django.template.loaders.cached.Loader', [
-                    'plugin.loader.PluginTemplateLoader',
+                    'plugin.template.PluginTemplateLoader',
                     'django.template.loaders.filesystem.Loader',
                     'django.template.loaders.app_directories.Loader',
                 ])
